requirements.txt
selenium
webdriver-manager
pyyaml
fake-useragent
pandas
google-genai
